refactor: moved interface class from demo project to main project
